Update Splits.py
A. Splits
time limit per test1 second
memory limit per test256 megabytes
inputstandard input
outputstandard output
Let's define a split of 
n
 as a nonincreasing sequence of positive integers, the sum of which is 
n
.

For example, the following sequences are splits of 
8
: 
[
4
,
4
]
, 
[
3
,
3
,
2
]
, 
[
2
,
2
,
1
,
1
,
1
,
1
]
, 
[
5
,
2
,
1
]
.

The following sequences aren't splits of 
8
: 
[
1
,
7
]
, 
[
5
,
4
]
, 
[
11
,
−
3
]
, 
[
1
,
1
,
4
,
1
,
1
]
.

The weight of a split is the number of elements in the split that are equal to the first element. For example, the weight of the split 
[
1
,
1
,
1
,
1
,
1
]
 is 
5
, the weight of the split 
[
5
,
5
,
3
,
3
,
3
]
 is 
2
 and the weight of the split 
[
9
]
 equals 
1
.

For a given 
n
, find out the number of different weights of its splits.

Input
The first line contains one integer 
n
 (
1
≤
n
≤
10
9
).

Output
Output one integer — the answer to the problem.